pyproject.toml
[tool.poetry]
name = "secret"
version = "0.1.0"
description = ""
authors = ["IDL13 <leonov.ilya.02@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
uvicorn = "^0.30.1"
pydantic = "^2.7.4"
pymongo = "^4.7.3"
pytest = "^8.2.2"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
